Add request body credential extraction
When using content negotiation the parsed body of a request is filled.
This eliminates the need to have a separate JSON extractor.

Fixes #9
